Update unit tests for OidcRpInitiatedLogout class
